# restful api/expApi/src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API da Loja virtual",
    "description": "Documentação da API",
    "version": "1.0.0"
  },
  "host": "localhost:1000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/v1/produto/": {
      "get": {
        "tags": [
          "Produto"
        ],
        "summary": "Lista todos os produtos.",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista de produtos retornada com sucesso."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "post": {
        "tags": [
          "Produto"
        ],
        "summary": "Adiciona um novo produto na base.",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "description": "Objeto contendo os dados do produto",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateProdutoDto"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Produto"
            },
            "description": "OK"
          }
        }
      }
    },
    "/v1/produto/{id}": {
      "get": {
        "tags": [
          "Produto"
        ],
        "summary": "Recupera dados de um produto específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do produto"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Produto"
            },
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Produto"
        ],
        "summary": "Atualiza dados de um produto específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do produto"
          },
          {
            "name": "body",
            "description": "Objeto contendo as alterações do produto",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateProdutoDto"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Atualização bem-sucedida."
          },
          "404": {
            "description": "Produto não encontrado."
          },
          "409": {
            "description": "Nome de produto já existente."
          }
        }
      },
      "delete": {
        "tags": [
          "Produto"
        ],
        "summary": "Remove um produto específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do produto"
          }
        ],
        "responses": {
          "204": {
            "description": "Remoção bem-sucedida."
          },
          "404": {
            "description": "Produto não encontrado."
          }
        }
      }
    }
  },
  "definitions": {
...
    "Produto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "8a2053de-5d92-4c43-97c0-c9b2b0d56703"
        },
        "nome": {
          "type": "string",
          "example": "Cadeira"
        },
        "preco": {
          "type": "number",
          "example": 15.99
        },
        "estoque": {
          "type": "number",
          "example": 10
        },
        "createdAt": {
          "type": "string",
          "example": "2024-04-26T17:19:15.645Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2023-04-26T17:19:15.645Z"
        }
      }
    },
    "CreateProdutoDto": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Cadeira"
        },
        "preco": {
          "type": "number",
          "example": 15.99
        },
        "estoque": {
          "type": "number",
          "example": 10
        }
      }
    },
    "UpdateProdutoDto": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Mesa"
        },
        "preco": {
          "type": "number",
          "example": 29.9
        },
        "estoque": {
          "type": "number",
          "example": 15
        }
      }
    },
...
  }
}